dom/tests/mochitest/chrome/test_geolocation.xul

<?xml version="1.0"?>
<?xml-stylesheet href="chrome://global/skin" type="text/css"?>
<?xml-stylesheet href="chrome://mochikit/content/tests/SimpleTest/test.css" type="text/css"?>
<!--
  Test for Geolocation in chrome
  -->
<window id="sample-window" width="400" height="400"
        xmlns="http://www.mozilla.org/keymaster/gatekeeper/there.is.only.xul">
  <script type="application/javascript" 
          src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>      

<script>
SimpleTest.waitForExplicitFinish();

var Ci = Components.interfaces;
var Cc = Components.classes;

var geolocation = Cc["@mozilla.org/geolocation;1"].getService(Ci.nsISupports);
geolocation.getCurrentPosition(done, error);

function error(error)
{
  ok(0, "error occured trying to get geolocation from chrome");
  SimpleTest.finish();
}
function done(position)
{
  ok(position, "geolocation was found from chrome");
  SimpleTest.finish();
}
</script>

<body xmlns="http://www.w3.org/1999/xhtml" style="height: 300px; overflow: auto;"/>

</window>
